# fix vector interpolation and remove unnecessary `@inbounds`