This is a followup to https://github.com/SciML/SciMLBase.jl/pull/476 that fixes some typos and also fixes a bounds error. This also removes unnecessary `@inbounds` that don't improve performance.

Benchmark:
```
julia> using Sundials

julia> mutable struct precflags
           prec_used::Bool
           psetup_used::Bool
       end

julia> prob = DAEProblem(prob_dae_resrob.f, prob_dae_resrob.du0, prob_dae_resrob.u0,
           prob_dae_resrob.tspan,  precflags(false, false));

julia> sol = solve(prob, IDA());

#before
julia> @btime sol(85000)
  3.474 μs (55 allocations: 4.67 KiB)

julia> @btime sol(1000:1000:90000);
  283.921 μs (4776 allocations: 420.81 KiB)

#after
julia> @btime sol(85000)
  3.450 μs (55 allocations: 4.67 KiB)
  
julia> @btime sol(1000:1000:90000);
  284.034 μs (4776 allocations: 420.81 KiB)
```